requirements.txt
Flask==3.0.0
google-cloud-vision==3.4.0
gunicorn==21.2.0
